refactor(py/genkit): genkit.ai -> genkit.blocks; genkit.veneer -> genkit.ai
RATIONALE:

- The primary top level namespace is shared across packages and crosses
  package boundaries. Doing this prevents us from defining any functionality
  within the top level namespace in an `__init__.py` file.
- The `genkit.veneer` namespace is kludgy at best and it makes more
  sense for users to use `genkit.ai` instead.
- Since the previously-called `genkit.ai` module is internal to Genkit
  and defines building blocks for various AI components, we've gone
  ahead and renamed it to `genkit.blocks`. We could call it anything
  else, but this seems like a reasonable name.

REFERENCES:

- [ ] https://packaging.python.org/en/latest/guides/packaging-namespace-packages/#native-namespace-packages

CHANGELOG:

- [ ] Renames and updated references for the following:
  - [ ] `genkit.ai` -> `genkit.blocks`
  - [ ] `genkit.veneer` -> `genkit.ai`
- [ ] Some formatting and import fixes.
