+++ py/packages/genkit/src/genkit/ai/__init__.py
@@ -1,28 +1,17 @@
 # Copyright 2025 Google LLC
 # SPDX-License-Identifier: Apache-2.0
 
+"""Veneer package for managing server and client interactions.
 
-"""AI foundations for the Genkit framework.
-
-This package provides the artificial intelligence and machine learning
-capabilities of the Genkit framework. It includes:
-
-    - Model interfaces for various AI models
-    - Prompt management and templating
-    - AI-specific utilities and helpers
-
-The AI package enables seamless integration of AI models and capabilities
-into applications built with Genkit.
+This package provides functionality for managing server-side operations,
+including server configuration, runtime management, and client-server
+communication protocols.
 """
 
+from genkit.ai.plugin import Plugin
+from genkit.ai.veneer import Genkit
 
-def package_name() -> str:
-    """Get the fully qualified package name.
-
-    Returns:
-        The string 'genkit.ai', which is the fully qualified package name.
-    """
-    return 'genkit.ai'
-
-
-__all__ = ['package_name']
+__all__ = [
+    'Genkit',
+    'Plugin',
+]
