+++ py/packages/genkit/src/genkit/blocks/document.py
