+++ py/packages/genkit/src/genkit/blocks/embedding.py
