+++ py/packages/genkit/tests/genkit/blocks/message_utils_test.py
@@ -5,7 +5,7 @@
 
 """Tests for the message utils."""
 
-from genkit.ai.messages import inject_instructions
+from genkit.blocks.messages import inject_instructions
 from genkit.core.typing import (
     Message,
     Metadata,
