+++ py/packages/genkit/tests/genkit/blocks/middleware_test.py
@@ -8,7 +8,7 @@ import asyncio
 
 import pytest
 
-from genkit.ai.middleware import augment_with_context
+from genkit.blocks.middleware import augment_with_context
 from genkit.core.action import ActionRunContext
 from genkit.core.typing import (
     DocumentData,
