+++ py/plugins/compat-oai/tests/test_plugin.py
@@ -3,11 +3,11 @@
 
 from unittest.mock import MagicMock, patch
 
+from genkit.ai.registry import GenkitRegistry
 from genkit.plugins.compat_oai.models.model_info import (
     SUPPORTED_OPENAI_MODELS,
 )
 from genkit.plugins.compat_oai.openai_plugin import OpenAI, openai_model
-from genkit.veneer.registry import GenkitRegistry
 
 
 def test_openai_plugin_initialize():
