+++ py/plugins/google-genai/src/genkit/plugins/google_genai/google.py
@@ -8,9 +8,9 @@ from google.auth.credentials import Credentials
 from google.genai.client import DebugConfig
 from google.genai.types import HttpOptions, HttpOptionsDict
 
+from genkit.ai.plugin import Plugin
+from genkit.ai.registry import GenkitRegistry
 from genkit.plugins.google_genai.models.gemini import GeminiModel, GeminiVersion
-from genkit.veneer.plugin import Plugin
-from genkit.veneer.registry import GenkitRegistry
 
 PLUGIN_NAME = 'google_genai'
