+++ py/plugins/vertex-ai/src/genkit/plugins/vertex_ai/plugin_api.py
@@ -8,12 +8,12 @@ import os
 
 import vertexai
 
+from genkit.ai.plugin import Plugin
+from genkit.ai.registry import GenkitRegistry
 from genkit.plugins.vertex_ai import constants as const
 from genkit.plugins.vertex_ai.embedding import Embedder, EmbeddingModels
 from genkit.plugins.vertex_ai.gemini import Gemini, GeminiVersion
 from genkit.plugins.vertex_ai.imagen import Imagen, ImagenVersion
-from genkit.veneer.plugin import Plugin
-from genkit.veneer.registry import GenkitRegistry
 
 LOG = logging.getLogger(__name__)
