+++ py/plugins/vertex-ai/tests/test_embeddings.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from genkit.ai.document import Document
+from genkit.blocks.document import Document
 from genkit.core.typing import EmbedRequest, EmbedResponse
 from genkit.plugins.vertex_ai.embedding import (
     Embedder,
