+++ py/samples/hello/src/hello.py
@@ -8,8 +8,9 @@ from typing import Any
 
 from pydantic import BaseModel, Field
 
-from genkit.ai.document import Document
-from genkit.ai.generate import generate_action
+from genkit.ai import Genkit
+from genkit.blocks.document import Document
+from genkit.blocks.generate import generate_action
 from genkit.core.action import ActionRunContext
 from genkit.core.typing import (
     GenerateActionOptions,
@@ -31,7 +32,6 @@ from genkit.plugins.vertex_ai import (
     VertexAI,
     vertexai_name,
 )
-from genkit.veneer.veneer import Genkit
 
 ai = Genkit(
     plugins=[VertexAI()],
