+++ py/tests/smoke/package_test.py
@@ -6,7 +6,7 @@
 # TODO: Replace this with proper imports once we have a proper implementation.
 from dotpromptz import package_name as dotpromptz_package_name
 
-from genkit.ai import package_name as ai_package_name
+from genkit.blocks import package_name as ai_package_name
 from genkit.core import package_name as core_package_name
 from genkit.plugins.chroma import package_name as chroma_package_name
 from genkit.plugins.firebase import package_name as firebase_package_name
